SCHED-615: Bug fixes and reordering.
