+++ scheduler/core/builder/validationbuilder.py
@@ -10,7 +10,7 @@ from lucupy.types import ZeroTime
 from .blueprint import CollectorBlueprint
 from .schedulerbuilder import SchedulerBuilder
 from scheduler.core.components.collector import Collector
-from scheduler.core.sources import Sources, Origins
+from scheduler.core.sources.sources import Sources, Origins
 from scheduler.core.programprovider.ocs import ocs_program_data, OcsProgramProvider
 from scheduler.core.statscalculator import StatCalculator
 from scheduler.core.eventsqueue import EventQueue
