+++ scheduler/core/plans/__init__.py
@@ -1,7 +1,7 @@
 # Copyright (c) 2016-2024 Association of Universities for Research in Astronomy, Inc. (AURA)
 # For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-from nightstats import *
-from plan import *
-from plans import *
-from visit import *
+from .nightstats import *
+from .plan import *
+from .plans import *
+from .visit import *
