+++ scheduler/core/plans/nightstats.py
@@ -1,11 +1,10 @@
 # Copyright (c) 2016-2024 Association of Universities for Research in Astronomy, Inc. (AURA)
 # For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-
 from lucupy.minimodel import Band, ProgramID
 
 from dataclasses import dataclass
-from typing import final, Mapping
+from typing import final, Dict
 
 __all__ = [
     'NightStats',
@@ -18,5 +17,5 @@ class NightStats:
     time_loss: str
     plan_score: float
     n_toos: int
-    completion_fraction: Mapping[Band, int]
-    program_completion: Mapping[ProgramID, str]
+    completion_fraction: Dict[Band, int]
+    program_completion: Dict[ProgramID, str]
