+++ scheduler/core/plans/plan.py
@@ -8,6 +8,9 @@ from typing import final, List, Optional, Tuple
 from lucupy.minimodel import Observation, ObservationID, Resource, Site
 from lucupy.timeutils import time2slots
 
+from .nightstats import NightStats
+from .visit import Visit
+
 
 __all__ = [
     'Plan',
