+++ scheduler/core/programprovider/ocs/__init__.py
@@ -1,5 +1,4 @@
 # Copyright (c) 2016-2024 Association of Universities for Research in Astronomy, Inc. (AURA)
 # For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-
-from ocsprogramprovider import *
+from .ocsprogramprovider import *
