+++ scheduler/core/sources/__init__.py
@@ -1,4 +1,8 @@
 # Copyright (c) 2016-2024 Association of Universities for Research in Astronomy, Inc. (AURA)
 # For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-from .statscalculator import *
+# TODO: This package will need significant cleanup after the initial demo version is released.
+
+from .origins import *
+from .services import *
+from .sources import *
