+++ tests/unit/scheduler/core/collector/collector_fixture.py
@@ -8,7 +8,7 @@ from lucupy.minimodel import ALL_SITES, NightIndex, Semester, SemesterHalf
 
 from scheduler.core.builder.blueprint import CollectorBlueprint
 from scheduler.core.builder.validationbuilder import ValidationBuilder
-from scheduler.core.sources import Sources
+from scheduler.core.sources.sources import Sources
 from scheduler.core.eventsqueue import EventQueue
