+++ tests/unit/scheduler/core/programprovider/ocs/test_ocs_filtering.py
@@ -8,7 +8,7 @@ import pytest
 from lucupy.minimodel import ObservationClass, Program
 
 from scheduler.core.programprovider.ocs import ocs_program_data, OcsProgramProvider
-from scheduler.core.sources import Sources
+from scheduler.core.sources.sources import Sources
 
 
 @pytest.fixture
